// web/src/api/model/feedbackResponse.ts
/**
 * Generated by orval v8.8.0 🍺
 * Do not edit manually.
 * Scholia API
 * OpenAPI spec version: 0.1.0
 */
import type { FeedbackHandler } from "./feedbackHandler";
import type { FeedbackStatus } from "./feedbackStatus";
import type { FeedbackSubmitter } from "./feedbackSubmitter";

export interface FeedbackResponse {
    /** @nullable */
    admin_notes?: string | null;
    body: string;
    created_at: string;
    handled_by?: null | FeedbackHandler;
    id: string;
    status: FeedbackStatus;
    submitter?: null | FeedbackSubmitter;
    updated_at: string;
    /** @nullable */
    url?: string | null;
    /** @nullable */
    user_agent?: string | null;
    /** @nullable */
    viewport_h?: number | null;
    /** @nullable */
    viewport_w?: number | null;
}